+++ lib/src/commands/commands.dart
@@ -1,9 +1,9 @@
+export 'add.dart';
 export 'bundle.dart';
 export 'cache.dart';
 export 'get.dart';
 export 'init.dart';
-export 'install.dart';
 export 'list.dart';
 export 'make.dart';
 export 'new.dart';
-export 'uninstall.dart';
+export 'remove.dart';