+++ test/command_runner_test.dart
@@ -1,15 +1,38 @@
 // ignore_for_file: no_adjacent_strings_in_list
 import 'package:args/command_runner.dart';
-import 'package:io/io.dart';
 import 'package:mason/mason.dart';
+import 'package:mason/src/command_runner.dart';
+import 'package:mason/src/io.dart';
+import 'package:mason/src/version.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
-import 'package:mason/src/command_runner.dart';
-import 'package:mason/src/version.dart';
 
 import 'helpers/helpers.dart';
 
 class MockLogger extends Mock implements Logger {}
+
+const expectedUsage = [
+  '⛏️  mason • lay the foundation!\n'
+      '\n'
+      'Usage: mason <command> [arguments]\n'
+      '\n'
+      'Global options:\n'
+      '-h, --help       Print this usage information.\n'
+      '    --version    Print the current version.\n'
+      '\n'
+      'Available commands:\n'
+      '  add      Adds a brick from a local or remote source.\n'
+      '  bundle   Generates a bundle from a brick template.\n'
+      '  cache    Interact with mason cache.\n'
+      '  get      Gets all bricks in the nearest mason.yaml.\n'
+      '  init     Initialize mason in the current directory.\n'
+      '  list     Lists all available bricks.\n'
+      '  make     Generate code using an existing brick template.\n'
+      '  new      Creates a new brick template.\n'
+      '  remove   Removes a brick.\n'
+      '\n'
+      'Run "mason help <command>" for more information about a command.'
+];
 
 void main() {
   group('MasonCommandRunner', () {
@@ -59,65 +82,21 @@ void main() {
       });
 
       test('handles no command', overridePrint(() async {
-        const expectedPrintLogs = [
-          '⛏️  mason • lay the foundation!\n'
-              '\n'
-              'Usage: mason <command> [arguments]\n'
-              '\n'
-              'Global options:\n'
-              '-h, --help       Print this usage information.\n'
-              '    --version    Print the current version.\n'
-              '\n'
-              'Available commands:\n'
-              '  bundle      Generates a bundle from a brick template.\n'
-              '  cache       Interact with mason cache.\n'
-              '  get         Gets all bricks.\n'
-              '  init        Initialize mason in the current directory.\n'
-              '  install     Installs a brick globally.\n'
-              '  list        Lists all available bricks.\n'
-              '  make        Generate code using an existing brick template.\n'
-              '  new         Creates a new brick template.\n'
-              '  uninstall   Uninstalls a brick globally.\n'
-              '\n'
-              'Run "mason help <command>" for more information about a command.'
-        ];
         final result = await commandRunner.run([]);
-        expect(printLogs, equals(expectedPrintLogs));
+        expect(printLogs, equals(expectedUsage));
         expect(result, equals(ExitCode.success.code));
       }));
 
       group('--help', () {
         test('outputs usage', overridePrint(() async {
-          const expectedPrintLogs = [
-            '⛏️  mason • lay the foundation!\n'
-                '\n'
-                'Usage: mason <command> [arguments]\n'
-                '\n'
-                'Global options:\n'
-                '-h, --help       Print this usage information.\n'
-                '    --version    Print the current version.\n'
-                '\n'
-                'Available commands:\n'
-                '  bundle      Generates a bundle from a brick template.\n'
-                '  cache       Interact with mason cache.\n'
-                '  get         Gets all bricks.\n'
-                '  init        Initialize mason in the current directory.\n'
-                '  install     Installs a brick globally.\n'
-                '  list        Lists all available bricks.\n'
-                '''  make        Generate code using an existing brick template.\n'''
-                '  new         Creates a new brick template.\n'
-                '  uninstall   Uninstalls a brick globally.\n'
-                '\n'
-                '''Run "mason help <command>" for more information about a command.'''
-          ];
           final result = await commandRunner.run(['--help']);
-          expect(printLogs, equals(expectedPrintLogs));
+          expect(printLogs, equals(expectedUsage));
           expect(result, equals(ExitCode.success.code));
 
           printLogs.clear();
 
           final resultAbbr = await commandRunner.run(['-h']);
-          expect(printLogs, equals(expectedPrintLogs));
+          expect(printLogs, equals(expectedUsage));
           expect(resultAbbr, equals(ExitCode.success.code));
         }));
       });